Renaming e criação de pastas para os componentes Vue do projeto.
