+++ equiv-frontend/src/components/StudentPage/ProcessEquivalences.vue
