+++ equiv-frontend/src/components/StudentPage/ProcessFiles.vue
@@ -40,7 +40,7 @@
 </template>
 
 <script>
-    import UserService from '../services/user.service';
+    import UserService from '../../services/user.service';
     export default {
         name: "ProcessFiles",
         data() {
