+++ equiv-frontend/src/views/LoginPage.vue
@@ -5,7 +5,7 @@
 </template>
 
 <script>
-    import Login from "../components/Login";
+    import Login from "../components/LoginPage/Login";
     export default {
         name: "LoginPage",
         components: {Login}
